Partially fix compilation of mCRL2 with clang.
Update aterm library header to include like:
 #include "mcrl2/aterm/some_header.h"
instead of
 #include "some_header.h"

The (tricky) situation that occurs is the following:
Building mCRL2 using clang requires a newer version of Boost. Having installed boost in $PREFIX automatically add $PREFIX/include to the includes passed to the compiler. Also having a muCRL installation installed in $PREFIX, has install the muCRL version of the aterm headers to $PREFIX/include. As during compilation, the include path for the boost headers is passed first, the compiler took the muCRL aterm library includes instead of the mCRL2 ones, causing all sorts of compile errors.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9274 21252fe5-60e8-0310-8ff0-c1ae1001594e