Added more quantifier tests. In particular one for forall x:Nat.x==3 which ought not to be true...
Adapted some of the enumeration code, to deal with more complex quantifications.

Moved some of the translation functions to internal format from the three
remaining rewriters to rewrite.cpp, and standardized these over the 
three rewriting engines. Actually, the internal formats were almost the
same, except that different strings were used.

The next step is to isolate the rest of the code that performs the
translation, esp. toInner and toInnerc, as well as fromInner.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9300 21252fe5-60e8-0310-8ff0-c1ae1001594e