Resolve error report #779 saying that { b: Bool | false}  == {} fails to be resolved properly:
The problem is that the rules for equality on sets cannot resolve this, because
it fails to determine that lambda x:D.false is equal to @false_  (as defined in set.spec).
But by replacing the rules

    ==(f, g)  ->  ==(@set(f, s), @set(g, t))  =  ==(s, t);
    !=(f, g)  ->  ==(@set(f, s), @set(g, t))  =  forall(c:S, ==(in(c, @set(f, s)), in(c, @set(g, t))));

by

==(@set(f, s), @set(g, t))  =  forall(c:S, ==(==(f(c),g(c)),==(@fset_in(c, s), @fset_in(c,t))));

equality of the emptyset is correctly established. This is now possible to the new rules for
resolving the universal quantifier.

A similar transformation is applied to equality of bags, but it is not clear whether
that will have undesired side effects.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9385 21252fe5-60e8-0310-8ff0-c1ae1001594e