Fix the order of compilation when using LTSmin integration.
The script did not work reliably, because sometimes the installation of
LTSmin was started before the mCRL2 libraries were installed. This lead to
non-compilation of the LTSmin tools that integrate with mCRL2.

The fix that was used is similar to the one proposed on
http://permalink.gmane.org/gmane.comp.programming.tools.cmake.user/37381

Note that when calling an external script, most of the CMake variables are
not available, hence a workaround has also been applied to pass on the flag
for shared vs. static building.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@9453 21252fe5-60e8-0310-8ff0-c1ae1001594e