Always add equalities to rewriter.
This prevents a tricky bug when using the compiling rewriter. It happened that, while rewriting expressions *inside the compiled rewriter*, quantifier elimination was applied. The first time quantifier elimination is applied, an enumerator is constructed. In the construction of the enumerator, all equations dealing with equality are rewritten. This could lead to the addition of new rewrite rules (for the equalities) being added to the rewriter, hence leading to compiling the rewriter anew. Compiling and reloading the rewriter, would however lead to the return addresses of the functions in the previous compiled rewriter to become invalid, causing a segmentation fault.

It seems this issue is not present in the code in trunk, but the difference between the two code bases is too large to accurately verify this.

Fixes the issue reported by Erik de Vink

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@9551 21252fe5-60e8-0310-8ff0-c1ae1001594e