- Removed parse_action_rename_spec, and use the new parser instead.
- Fixed the action rename parser.

- Made action rename data structures more conforming to the rest of the mCRL2 libraries.

  * Changed the attributes of action_rename_rule and action_rename_rule_rhs such that they can contain an incomplete object (after type checking).
  
  * Added to_aterm conversion functions.
  
  * Fixed some errors in the interfaces (returning by value instead of by const reference).

The action rename test now fails. I don't know yet if this is caused by an error in the parser or by side effects of the action rename interfaces that are no longer available (in particular the error-prone interface of action_rename_rule_rhs, which is misused in at least one place in the action rename algorithm).


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9812 21252fe5-60e8-0310-8ff0-c1ae1001594e