Added a formalisation of (a part of) the DIRAC (Distributed Infrastructure with Remote Agent Control) system. This is the Grid solution designed to support production activities as well as user data analysis for the LHC-beauty experiment at CERN. The system consists of cooperating distributed services and a plethora of light-weight agents delivering the workload to the Grid resources.
Two subsystems of the DIRAC systems have been modelled.  Despite the
10+ years of continuous improvement of the DIRAC system, the analysis
of the formal models of two subsystems reveals several problems of the
system, such as the presence of livelocks and race conditions, which
were confirmed to occur in the real system.

The formal models also make for some good benchmark material for state
space exploration: with the current mCRL2 state space exploration
techniques, generating a state space for WMS.mcrl2 takes approx. 50
hours on a contemporary workstation. This contrasts starkly to the ca. 10
seconds required to perform a symbolic exploration using the lps-reach
tool of LTSmin.

Joint work with 
* Daniela Remenska and Jeff Templon (Nikhef), 
* Henri Bal, Kees Verstoep and Wan Fokkink (VU Amsterdam)
* Tim Willemse (TU/e)




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10065 21252fe5-60e8-0310-8ff0-c1ae1001594e