Fix generation of backtrace in ltsview/simulation.cpp.
The algorithm for generating a backtrace always considered the first incoming edge of a state. This could however be a self-loop or a back-edge. A self-loop will always lead to an infinite loop, and back-edge will never lead to the shortest backtrace, and may lead to an infinite loop.

Note that the only state in the system that can be such that all its incoming edges are self-loops or back-edges can be the initial state; all other states have at least one incoming edge that is not a self-loop and not a back-edge.
This fix makes sure that the first such edge is used.

fixes: #861

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10172 21252fe5-60e8-0310-8ff0-c1ae1001594e