In the cleanup of the garbage collection code, I exchanged initializing first_free and AT_allocate, leading to the loss of one free AFun each time garbage collection took place.
This has now been repaired, by changing the order back.

It appears that the clang compiler even assigned wrong
values due to this problem, most likely due to code optimisation, 
and reallocating vectors. At least the crashes that occurred
under the clang compiler seem to have gone away, which
could mean that this is resolving bug #1007. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10438 21252fe5-60e8-0310-8ff0-c1ae1001594e