Increased the tendency to collect garbage instead of increasing the ATerm hash table to reduce the memory footprint from a factor 4 to a factor 8.
Changing it from a factor 2 to 4 gave a substantial reduction
in memory usage in some cases (see the performance benchmarks).

Depending on the effect, this factor 8 can even be increased.
Contrary to the previous garbage collection strategy, the current
one is amortized constant divided over the number of new terms
that are being added in each round. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@10440 21252fe5-60e8-0310-8ff0-c1ae1001594e