Effectively replaced ATermAppl and ATermList, as well as atermpp::term_appl<T> and atermpp::term_list<T> by aterm::term_appl<T> and aterm::term_list<T>. Removed functions in classes that were not strictly necessary, or obsolete. Adapted the toolset to these new changes.
All tests succeed, including the balanced_tree_test.

The next steps are to adapt the sorts ATermTable and ATermIndexedSet
to become proper classes, conforming the standard set in the atermpp 
library.

If that is done, the aterm classes must be moved to the atermpp 
library, and the AT... functions must become derived functions
from this class.

Finally, performance must be addressed. Currently, the issue of
performance is totally ignored, as all energy goes into keeping
the system correct.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@10547 21252fe5-60e8-0310-8ff0-c1ae1001594e