Took care that the experimental tools also compile properly with the new aterm implementation.
Investigated the problems with the tests on Linux. All tests now succeed
on linux and on the mac. Problems were most likely caused due to a different
order in which data structures are being removed on the mac under clang, 
compared to the way that is done in linux with g++.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@10601 21252fe5-60e8-0310-8ff0-c1ae1001594e