Fix awful bug in static initialisation order.
In atermpp/aterm.h, a hashtable was defined as
  extern std::vector <detail::_aterm*> hashtable;
It was defined in atermpp/memory.cpp.

In atermpp/aterm_list.h empty_list() is defined statically as well; the
initialisation of this static uses aterm(AS_EMPTY_LIST), which, in its
definition uses the hashtable. We do not know in which compilation unit
empty_list() is compiled, but we can be pretty sure that it won't be
atermpp/memory.cpp.

For initialisation of static variables in *different* compilation units, c++
does not guarantee anything about the order in which initialisation will occur.

This is now fixed by using:
  extern std::vector<detail::_aterm*>& hashtable;
i.e. we wrap the static initialisation in a function to ensure a proper
initialisation order.

For more information see http://www.parashift.com/c++-faq-lite/ctors.html#faq-10.14,
which has a nice explanation on this problem.

Note that things could still go wrong if some of the destructors of static
objects also use the hashtable. If that is the case, then we should wrap
a pointer to the hashtable instead.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@10682 21252fe5-60e8-0310-8ff0-c1ae1001594e