Resolves bug #1024.
A .aut file with non consecutive state numbers is now
read. This means that a file written using bit hashing can now be read.

The rational test has been adapted. The ltsconvert tests should not be
run, because lps2lts does not work. This does not work with -rjittyp. 
As lps2lts does not work not .aut, .fsm and .lts files are generated.
This means ltsinfo, ltsconvert and the ltscompare tests can not be
run. For reasons that I do not understand the ltscompare tests are still
running on my machine.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11064 21252fe5-60e8-0310-8ff0-c1ae1001594e