Merge r11064 and r11065 from trunk
r11064:
Resolves bug #1024.

A .aut file with non consecutive state numbers is now
read. This means that a file written using bit hashing can now be read.

The rational test has been adapted. The ltsconvert tests should not be
run, because lps2lts does not work. This does not work with -rjittyp.
As lps2lts does not work not .aut, .fsm and .lts files are generated.
This means ltsinfo, ltsconvert and the ltscompare tests can not be
run. For reasons that I do not understand the ltscompare tests are still
running on my machine.

r11065:
I left two residual errors, one due to me in the previous commit.

When reading a .aut file the initial state and the value 0 were
mixed up in the translation table.

When writing a .aut file generated using bit hashing, the initial
state was set to 0, instead of to the bit hashed valued. This
led to a transition system with one state too many, and unreachable
transitions.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11100 21252fe5-60e8-0310-8ff0-c1ae1001594e