Merge r11145, r11146, r11147 from trunk
r11145:
Clean up test for ltscompare.

r11146:
Add test case for bug #1082

r11147:
Treat function symbol for trace_pair in standard way.

The handling of trace_pairs, which was done all manually, lead to crashes. We now use the same mechanism of static initialisation that is used in other places as well.

Fix #1082

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11153 21252fe5-60e8-0310-8ff0-c1ae1001594e