Merge r11171 and r11172 from trunk
r11171:
Attempt to fix bundle creation on MacOSX.

This reintroduces the fix from r10359, which was inadvertedly removed in r10954.
For creating bundles on MacOSX we need to ensure that the boost libraries and the header files have been packaged into the bundle, in order to build the compiling rewriters.

This needs testing.

r11172:
Fix changeset in r11171

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11187 21252fe5-60e8-0310-8ff0-c1ae1001594e