Merge r11178 from trunk
Repaired an obvious error in jittyc, where
due to a typing mistake the "or" case was not treated
properly. This error may have been in the jittyc rewriter
for years. It is not clear why it was not triggered earlier.
It was found in the toolset based on the new aterm library,
where it caused failing tests.

Updated the CMakeLists file for lpsrealelm, to allow
printing of summands.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@11189 21252fe5-60e8-0310-8ff0-c1ae1001594e