[release] Properly document changeset r11180, and clean it up.
The change to jittyc.cpp that was made in r11180 could be done in a lot more readable way, by (re)using the file_exists function, that was implemented in uncompiledlibrary.h; I moved this to the more suitable file_utility.h.

Furthermore, this changeset adds clear documentation as to why these changes were implemented in this way (MacOSX), such that they are not likely to be removed inadvertedly in the future.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11194 21252fe5-60e8-0310-8ff0-c1ae1001594e