Reorganised reading and writing of aterms, by grouping all relevant code into aterm_io.h, aterm_io_text.cpp and aterm_io_binary.cpp.
Initialisation of the aterm library consisted of setting
appropriate io flags for windows. This has now been moved
to the io routines, and is only done when terms are read
or written. This means that explicit initialisation, of the 
aterm library which occured in many files, could now
be removed. 

Reading and writing of aterms still consists of lots of
duplicated code. Moreover, the io is file and string based.
It would be better if io would be stream based. This still
needs to be done.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@11332 21252fe5-60e8-0310-8ff0-c1ae1001594e