Fixes #1114.
This problem was caused by a wrong implementation/definition of the inverse of
a renaming operator (which was treated as if it were a bijection). The problem already 
shows up on the following simple example:

act c,d;
init
   allow({c}, rename({d -> c},  c));

The problem has been fixed, awaiting the new implementation of alphabet conversions.

It might be that more random testing will reveal other bugs in the current 
implementation of the alphabet reductions.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11362 21252fe5-60e8-0310-8ff0-c1ae1001594e