Resolved an aterm-leak in the svc library, detected by the strenghtened reference count check.
Made the reference count check a little bit stronger. Prevent the check
from stopping at the detection of the first problem, as more information
helps often to understand what is going on.

Removed spaces in the strings of the four basic symbols.

Only the trace_read_mcrl2 test now fails. This needs to be investigated further.

The lps_ltsmin_test also fails, but only when run in combination with
other tests. It succeeds when run standalonely (<-- English?). 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@11382 21252fe5-60e8-0310-8ff0-c1ae1001594e