Re-enabled the check that all reference counts of aterms are reset to 0 at the end of program run. All reference counts appear to be reset properly, except for the reference counts of
aterm static_undefined_aterm;

which is also used in the default constructor aterm() (and its
derivatives). On the mac these counts go back to zero at the
end of a program run, but on linux these can become very high.

Therefore, checking whether the reference count of static_undefined_aterm
is disabled, awaiting more understanding of what is going wrong.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@11390 21252fe5-60e8-0310-8ff0-c1ae1001594e