Adapted printing of strings such that they are always printed without quotes, unless they contain quotes, round or square brackets, \, start with - or a digit, a comma, tab, carriage return or newline. In the last case the string is printed with quotes, and the \ and ", as well as a tab, newline and CR are escaped by adding an extra \ in front of it.
Tests have been adapted to this changed form of printing.

A formula is added to the game "four in a row" to indicate
that red has a winning strategy. On a seven by six board,
which is the standard format, red has a winning strategy
as shown in 1988. Expectation is that this yields
a state space of 10Gstate, based on extrapolation with
boards of size k by six, with k smaller than 7. It appears
that in those cases red (as the first player) has no
winning strategy.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@11450 21252fe5-60e8-0310-8ff0-c1ae1001594e