Add an elegant iterator interface to applications.
As a preparation for moving to another representation of applications
in the data library, it was decided to add an iterator interface.

For the time being, we can simply proxy the iterators to the iterators
on the argument list. Note that we only allow const_iterators, and we explicitly
declare iterators as private to prevent access to the iterators that we
inherit from atermpp::aterm_appl.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/aterm@11517 21252fe5-60e8-0310-8ff0-c1ae1001594e