removed all gsIs... and gsMake... functions that can be removed.
There are a number of functions that are not removable, such
as gsIsNil and gsMakeNil, gsIsParamId, gsMakeParamId, gsMakeIdAssignment...
as they have no pendant in generated data type libraries.

In the rewriters and pbes these functions cannot be removed, as these
work in internal format.

And there are still a few more cases that cannot be eliminated
for various reasons.

All tests succeed on the mac.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11659 21252fe5-60e8-0310-8ff0-c1ae1001594e