Fix resetting in global analysis of stategraph.
In resetting, we need to check whether the control flow location that we are
considering is actually relevant in the recursion at hand. It can be
irrelevant if the destination of a control flow parameter is defined
for the recursion, and different from the value of the same control
flow parameter at the control flow location.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11683 21252fe5-60e8-0310-8ff0-c1ae1001594e