Resolves #1115.
The mutable_indexed_substitution has been given a set that
maintains the set of variables occurring in the righthandside of
variables. This set is only maintained if this set is once requested in
the set. With this set the Controllers example that became very slow
has reasonable performance again.

Jeroen: can you check the code? It seems to me that the clear
function on a mutable_indexed_substitution was incorrect as
it did not reset all the variables in the class.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11909 21252fe5-60e8-0310-8ff0-c1ae1001594e