Reverted a change to avoid case expressions of the form
  C(v,t,dc1,dc2,....,dcn)

where t is a term and dc_i are global or don't care variables. 
As an optimisation to accommodate constant elimination these expressions
were replaced by the simpler term t. But this in certain cases lead to
bigger state spaces. In particular it make the non reduced state space
for the lineariser slightly bigger but quite unreadable. 

The problem that constant elimination cannot deal properly with such expressions
must probably be dealt with in the algorithms for constant elimination.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@11952 21252fe5-60e8-0310-8ff0-c1ae1001594e