This resolves but #1210.
Clustering would map a variable in a sum sometimes onto a 
process parameter, in the rare case that there are at least
two sum operators with variables of the same sort, with different
name, of which one matches a process parameter.

This did not happen before, because the names of parameters in
sums could not match process parameters. Most likely due to the
introduction of the assignment syntax this is now possible.

The problem is repaired and Jeroen Keiren checked a testcase in,
just before this commit.

The problem was detected by Jeroen Redegeld, modelling XRay 
equipment at Philips.

Bugfix has been reviewed by Jeroen Keiren.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12092 21252fe5-60e8-0310-8ff0-c1ae1001594e