Merge r12088 to r12095 from trunk.
r12088:
- Improved error messages for ATerm file format version errors.

- Introduced some exceptions for ATerm I/O.

Patch reviewed by Jeroen.

r12089:
Disable lps2lts and pbes2bool tests for sets_bags.

These test cases timeout in a Maintainer build on most platforms; especially
lps2lts -yno, and
pbes2bool -c are the commands that cause timeouts; the first due to a large
state space because of the presence of global variables; the second due to a
large output.

r12090:
Remove another few terribly long running test cases (sets_bags)

r12091:
Add test case for a problem in the lineariser.

The problem was observed by Jeroen Redegeld, and analysed by JFG. A fix has been proposed, and this is the test that goes with it.

This adds a new type of test to the LPS library, in which a specification can
be linearised using a number of strategies; for the result an LTS is generated,
which is compared w.r.t. bisimilarity to an expected state space.

r12092:
This resolves but #1210.

Clustering would map a variable in a sum sometimes onto a
process parameter, in the rare case that there are at least
two sum operators with variables of the same sort, with different
name, of which one matches a process parameter.

This did not happen before, because the names of parameters in
sums could not match process parameters. Most likely due to the
introduction of the assignment syntax this is now possible.

The problem is repaired and Jeroen Keiren checked a testcase in,
just before this commit.

The problem was detected by Jeroen Redegeld, modelling XRay
equipment at Philips.

Bugfix has been reviewed by Jeroen Keiren.

r12093:
- Suppressed some unused variable warnings (noted by Jan Friso).

r12094:
Put atermpp library in the appropriate component for installation

r12095:
Add notes on regression tests in Jenkins, and its setup.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/release@12096 21252fe5-60e8-0310-8ff0-c1ae1001594e