Adapted the registration of hooks such that this is done without explicit initialisation. Rewriting appears to work properly using this numbering scheme.
Based the jitty and jittyc rewriter completely on the numbers 
generated by function symbol traits. Removed the old Muckian
scheme. Translation to innermost and outermost format now
consists of the identity functions.

There is a nasty problem which I do not completely understand
with an application constructor that takes a function to be
applied on the arguments before the application is made. This 
causes the adminstration of terms in the hashtable to go 
astray. This needs further investigation.

Lots of tests fail, most of them due to the problem
with recursive term construction and hastables.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/rewriter@12233 21252fe5-60e8-0310-8ff0-c1ae1001594e