Repaired a few problems that caused release tests to fail in the jittyc/jittyp rewriter.
The jittyp rewriter is becoming sensitive to abstractions (lambda, forall and exists quantifier)
and to where_clauses. In the past these were apparently silently ignored.

Started making the jittyc rewriter to deal with higher order terms.
Also dealt with a problem where new function symbols are added to the data specification,
without rebuilding the rewriter.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/rewriter@12271 21252fe5-60e8-0310-8ff0-c1ae1001594e