Carried out various optimisations, including some declarations from struct_core to a .cpp file.
The effect is that the compiling rewriter is exactly as fast
generating a state space for 1394-fin.lps as it was before
removing the internal format.

There are a number of additional performance improvements to the
rewriter that should be reinstalled. This means that it is not
unreasonable to expect that the new rewriter will outperform
the old one.

It is questionable whether moving declarations of gsOpId to a .cpp
file are safe. All tests succeed on the mac, but this may not
be the case on other platforms. Moreover, these are changes made
in generated code, so this need further investigation.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/rewriter@12292 21252fe5-60e8-0310-8ff0-c1ae1001594e