Tried to implement a suggestion for reliable static initialisation. This is applied to function_symbol and variable. All data is moved from struct_core into the class. So, the class function_symbol now contains a static member with the OpId atermpp::function_symbol.
This function symbol is initialised in data.cpp. 

Removed a few occurrences of auto_buffers to improve performance.

All tests succeed on a mac. It is not clear whether initialisation
in this way is safe. This needs to be discussed further. But as it
stands the current approach could be a good and efficient candidate.

Performance of generating the state space of 1394-fin slightly outperforms
the rewriter in the trunk. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/rewriter@12297 21252fe5-60e8-0310-8ff0-c1ae1001594e