Apply the final changes made to pbespgsolve by Maks Verver.
Essentially this applies Maks' patch sent to the mcrl2-users mailing list
on 28 Nov 2013, with the following description:

Attached is a patch to bring the pbespgsolve source up to date with the
latest version of the solver tool I created for my master's project.
For end users, the most notable changes should be some performance
improvements, both for the small progress measures and recursive
solvers.

Apart from applying this patch, it may still be desirable to:

- Tweak the default configuration.  Currently, pbespgsolve defaults to
Small Progress Measures as the solving algorithm with no preprocessing
operations.  Based on my research, the recursive algorithm with SCC
decomposition is probably a better selection.

- Remove some of the unused source files from the mCRL2 repository.
Currently the pbespgsolve directory contains a lot of files related to
features that I experimented with but which didn't work as well as I
hoped (multithreaded and distributed versions of the algorithms, for
example, as well as experimental lifting strategies).

I have applied this patch and made some essential changes to at least compile
warning-free on OS X using Clang. Some additional changes and clean up are
probably still needed for other platforms. Furthermore I have not actively
taken up the suggested changes mentioned above.

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12313 21252fe5-60e8-0310-8ff0-c1ae1001594e