Resolved #1260
The problem was that terms with a function update
did not rewrite properly.

The problem is that the function symbol for a function_update was not added
to the mappings (but the equations were). This is a rather old error.
The new rewriter looks into 
the function symbols in the specification to generate code for the compiler. 
As the function_symbol was not there, no code was generated.
The old rewriter would search the input and use the function symbols in the input
to generate the code for the rewriter. So, it overlooked that a function symbol
was missing in the data specification.

All tests succeed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12397 21252fe5-60e8-0310-8ff0-c1ae1001594e