Using CMake dependency mechanisms to link in extra libraries and add extra include directories. Sadly, this requires us to compile empty .cpp files for header-only libraries, but it still gives us cleaner CMake code than we would have otherwise. In particular, the build script for a tool like mcrl22lps now reads:
   add_mcrl2_tool(mcrl22lps SOURCES mcrl22lps.cpp DEPENDS mcrl2_lps)

A strange issue remains with tools that use the PBES library: although the BES library has a
link-time dependency on the PBES library, any code that uses the I/O functionality in the
PBES library creates a link-time dependency from the PBES library to the BES library again.
This can now only be resolved by including the dependencies in a particular order (in fact,
it results in the linker linking mcrl2_bes, then mcrl2_pbes, and then mcrl2_bes again). The
real conclusion here is that, as it stands, the PBES library and the BES library are not
separate libraries at the moment.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/qt5@12434 21252fe5-60e8-0310-8ff0-c1ae1001594e