Removed the use of pbes_expression_substitute_and_rewrite in favour of the quantor_eliminating_pbes_rewriter. There is one other occurrence of pbes_expression_substitute_and_rewrite, which also seems to be replaceable, given that the internal format is now equal to the external format in the toolset.
The trigger to do this, was a very subtle error in pbes_expression_substitute_and_rewrite
which when eliminating quantifiers over variables over sorts without constructors, 
would yield an expression with variables that were free but ought to be bound.

The advantage of this reduction is that substantial parts of redundant code can
be removed.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12523 21252fe5-60e8-0310-8ff0-c1ae1001594e