- Make the extra local control flow graph sparser. (patch supplied by Jeroen)
Only introduce an edge if there is a corresponding data parameter that belongs to
the extra graph. This corresponds to the following update of the pseudo code.

ComputeExtraCFG (E, (V_1,E_1,B_1),...,(V_n,E_n,B_n))
=
  V := emptyset;
  E := emptyset;
  B := emptyset;
  for X in bnd(E) do
    V := V union {(X,?,?=?)};
    B := {(X,d) | d is a data parameter of X };
    for i in 1 .. n do B := B setminus B_i;
    for j in 1.. npred(phi_X) do
      if exists k: d_X[k] \in B && (used(X, j, k) || changed(X, j, k)) then
        E := E union { ( (X,?,?=?),j, (pred(phi_X,j),?,?=?) ) };

Rationale:
We need to incorporate exactly this edges such that, if a parameter that belongs to this extra
graph is used or changed, then there also in an edge in the extra graph. As a consequence, this
graph has the property that with every use/change of the data parameter, the corresponding
PVI is ruled by the extra graph.

Formally:
Given PBES E and a collection of CFGs G_i = (V_i,E_i) with belongs-to relation B_i, for 1 <= i <= n.

Define extra graph (V,E) with belongs-to relation B as follows:
* V = bnd{E}

* for data variables d: (X,d) \in B(X) iff for all i: (X,d) \notin B_i

* E is a subset of V x Nat x V, defined as follows:
(X,j,Y) in E iff:
  1. j <= npred(phi_X) and,
  2. Y = Predvarname(phi_X, j), and
  3. exists (X,d) \in B(X): (X,d) is used for PVI(phi_X, j) or (X, d) is changed by PVI(phi_X, j)


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12620 21252fe5-60e8-0310-8ff0-c1ae1001594e