Rerouted all the use of EnumeratorSolutionsStandard through the classic enumerator. This allows to remove the class EnumeratorSolutionsStandard.
Temporarily added 

    /// \brief The rewriter needs to find fresh names for
    //         variables. This name generator can be used
    //         for other renaming purposes as well.
    data::set_identifier_generator& identifier_generator()
    {
      return m_rewriter->identifier_generator();
    }

to the interface of a basic rewriter. This is probably
not desired, and may need to be discussed and adapted, but for the time
being this was the only way that I could find to circumvent
EnumeraterSolutionStandard. Wrapping the rewriter class does not
work, because a similar wrap is required for the Rewriter class,
but as Rewriters are stored by shared pointers in the rewriter class,
this leads to undesired removal of the Rewriter class. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12689 21252fe5-60e8-0310-8ff0-c1ae1001594e