Adapted the interface of the classic enumerator to match the interface of Wieger's enumerator. The classic enumerator now employs enumerator_list_element_with_substitution. Some changes to this class were necessary. Wieger: can you check my changes.
This operation has not yet been finished. The interfaces are not yet equal.

All tests succeed.

Remarkably the state space generation of 1394-fin.lps turns out
to be 25% faster using the jittyc rewriter. Generation of this
state space is now faster than it has ever been. Apparently, the
classic enumerator was not so efficient after all, probably because
it copied too much information.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12787 21252fe5-60e8-0310-8ff0-c1ae1001594e