Replaced a large number of "atermpp::convert" by a more direct conversion. In many cases the convert could be ommitted completely. In other cases, it was replaced by typical code of the shape
    term_list(x.begin(),x.end())

instead of atermpp::convert<term_list>(x). The motivation for
this is that I believe that with the new aterm library, conversion
can be done directly, and now convert function is needed anymore
in the aterm library. 

I left one occurrence of convert in what I believe is generated
code of the parser. I tried to avoid replacing converts in generated
code.

All tests succeed.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12808 21252fe5-60e8-0310-8ff0-c1ae1001594e