Repaired a nasty error in indexed sets, that would only occur when elements were removed and reinserted.
Introduced three levels of erasing non used variables
in pbes2bool and pbes2bes. If all non used variables are
removed, reductions in memory usage of a factor 20 have
been observed. In combination with the depth-first option
this might even be more, but it may requires extra time
to recalculate the equations for removed variables.

Adapted the test cases to the change in erasure levels.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@12999 21252fe5-60e8-0310-8ff0-c1ae1001594e