[Release]
This commit addresses #1301, showing a typecheck issue in modal formulas,
using FSets, Sets and type aliases. The problem was reported by Willem Sonke.

The problem is due to the fact that in a modal formula the types
were not properly unwound to their basic form. So, CPUs was not replaced
by Set(CPU), or even more precise, Set(struct p1). Therefore, the routine
trying to upcast FSet(CPU) to Set(CPU), had to upcast FSet(CPU) to CPUs.
The latter failed.

The problem is resolved by unwinding the types always before applying
upcasting.

A testcase showing the problem has been added.

This solution is ad-hoc, awaiting the replacement of the typechecker by a new one.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13013 21252fe5-60e8-0310-8ff0-c1ae1001594e