[NOT FOR RELEASE]
Added the option to pbes2bool to limit the number of bes variables
that it needs to explore. This is the pendant of high-way search, which
can be enabled in lps2lts also using the --todo-max flag. 
If there are too many variables that must
be explored simultaneously, the rhs of these variables is simply set
to false (or true depending on whether the approximate-false flag is
used). This provides an approximation of the validity of the formula.

Using it peg-solitaire can be solved in 15 seconds (i.e. <true*><ready>true)
can be proven to hold, which takes appr. 100s with depth first search
and -s2. At other examples, such as four-in-a-row no real impressive
systematic speedup appears to be possible.





git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13018 21252fe5-60e8-0310-8ff0-c1ae1001594e