Proper probabilistic finite state automata are now generated in .aut format. All tests succeed. The generated .aut format cannot be read or processed yet.
As an example consider the following stochastic process in MCRL2:

act a:Bool;
proc X=dist p:Bool[1/2].a(p).dist p:Bool[if(p,1/4,3/4)].a(p).X;
init X;

The generated state space is

des (1 1/2 0,4,4)
(0,"a(false)",3 1/4 2)
(1,"a(true)",3 1/4 2)
(2,"a(false)",1 1/2 0)
(3,"a(true)",1 1/2 0)

Instead of an initial state an initial distribution is generated. 
Furthermore, at the right hand of the transition state distributions
are generated also.

A state distribution must be read as follows:

1 1/2 0

means that state 1 can be reached with probability 1/2 and 0 with
the remaining probability.

1 1/4 2 1/8 3 1/16 4

would mean that state 1 can be reached with probability 1/4, 
2 with 1/8, 3 with 1/16 and state 4 with the remaining 9/16 
probability.




git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13188 21252fe5-60e8-0310-8ff0-c1ae1001594e