Replaced the rewrite routines such that the functions rewr_... do not recall which arguments are normal forms, but this is now being recalled in the arguments of the rewrite functions. More precisely these arguments recall how they can be transformed to normal form in the most efficient way. This has a substantial effect on the time rewriting is taking in some cases. For instance generating the whole state space for the game of the goose for three players is more than 4 times faster.
The work is not completed, in the sense that a careful scrutiny is required to
see whether unnecesary rewriting is avoided everywhere. Especially, this is not
yet the case for higher order functions.

However, a bug that was recently introduced for higher order rewriting 
by revamping the jittyc rewriter has been repaired. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@13493 21252fe5-60e8-0310-8ff0-c1ae1001594e